git-svn-id: svn://123.57.173.36/meijia-parent@1761 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-service/src/main/java/com/simi/service/impl/async/UserScoreAsyncServiceImpl.java
@@ -94,7 +94,7 @@ public class UserScoreAsyncServiceImpl implements UserScoreAsyncService {
 		
 		CompanyAdminSearchVo searchVo = new CompanyAdminSearchVo();
 		searchVo.setUserId(userId);
-		searchVo.setIsCreate(1);
+		searchVo.setIsCreate((short) 1);
 		searchVo.setStartTime(startTime);
 		searchVo.setEndTime(endTime);
